# openapi.json
{"openapi":"3.1.0","info":{"title":"Social Media Microblog API","description":"A Twitter-like microblog API built with FastAPI and SQLite. Supports creating posts (max 280 chars) and liking posts (once per user).","contact":{"name":"Microblog API","email":"dev@microblog.com"},"version":"1.0.0"},"paths":{"/posts/":{"post":{"tags":["posts"],"summary":"Create a new post","description":"Create a new microblog post. Content must be between 1 and 280 characters.","operationId":"create_post_posts__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["posts"],"summary":"Get all posts","description":"Retrieve all posts ordered by newest first, with like counts.","operationId":"get_posts_posts__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"},"title":"Response Get Posts Posts  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/posts/{post_id}/like":{"post":{"tags":["posts"],"summary":"Like a post","description":"Like a post. Each user can only like a post once. Returns 409 on duplicate.","operationId":"like_post_posts__post_id__like_post","parameters":[{"name":"post_id","in":"path","required":true,"schema":{"type":"integer","title":"Post Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikeRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["health"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LikeRequest":{"properties":{"user_name":{"type":"string","title":"User Name"}},"type":"object","required":["user_name"],"title":"LikeRequest"},"LikeResponse":{"properties":{"message":{"type":"string","title":"Message"},"total_likes":{"type":"integer","title":"Total Likes"}},"type":"object","required":["message","total_likes"],"title":"LikeResponse"},"PostCreate":{"properties":{"content":{"type":"string","title":"Content"},"user_name":{"type":"string","title":"User Name"}},"type":"object","required":["content","user_name"],"title":"PostCreate"},"PostResponse":{"properties":{"id":{"type":"integer","title":"Id"},"content":{"type":"string","title":"Content"},"user_name":{"type":"string","title":"User Name"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"likes_count":{"type":"integer","title":"Likes Count"}},"type":"object","required":["id","content","user_name","created_at","likes_count"],"title":"PostResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}